requirements.txt
Django==4.2.*
djangorestframework==3.15.1
djoser==2.2.2
djangorestframework-simplejwt==5.3.1
django-cryptography===1.1
django-user-agents==0.4.0
celery==5.3.6
gevent==24.2.1
django-redis==5.4.0
flower==2.0.1
